Auto-hiding arrows, dot pagination, simpler usage
This changeset seems large but includes a “dotted” mechanism for
pagination, arrows that auto-hide (for first/last pages) and requires a
simpler underlying structure from a user perspective i.e. instead of
calling the plugin on a <div> that contains a <ul>, just call it on the
<ul> directly… the plugin will wrap it in a <div> which is required to
contain the navigation elements.  additionally, rather than
manipulating element styles directly, this version embeds a style-sheet
that gets attached to the document head that contains minimum styling
necessary for things to work (it may not be the best approach but I’m
open to a discussion about it)